move lgUtils to Shell
